Post page UI - camera implementation
