+++ .dart_tool/extension_discovery/devtools.json
@@ -1,0 +1,1 @@
+{"version":2,"entries":[{"package":"rewarding_sale_app_flutter_app","rootUri":"../","packageUri":"lib/"}]}
